requirements.txt
requests
mitmproxy
discordpy